put auto migration in common code
